KIBON-98 - Linkdialog: Dialog mit einem Link. Momentan im testdaten benutzt. Das alte angularjs-Dialog wurde entfernt, da es nicht mehr gebraucht wird. - DvDialog konnte aus ajs-upgrade-providers entfernt werden - In testDaten funktioniert alles nun mit MatDialog und nicht mehr mit den alten Dialogs von angularjs

+++ ebegu-web/src/admin/component/testdatenView/testdatenView.ts
@@ -14,15 +14,17 @@
  */
 
 import {Component, OnInit} from '@angular/core';
+import {MatDialog, MatDialogConfig} from '@angular/material';
 import {IPromise} from 'angular';
 import * as moment from 'moment';
-import {DvDialog} from '../../../core/directive/dv-dialog/dv-dialog';
+import {Observable} from 'rxjs/Observable';
+import {DvNgLinkDialogComponent} from '../../../core/component/dv-ng-link-dialog/dv-ng-link-dialog.component';
+import {DvNgOkDialogComponent} from '../../../core/component/dv-ng-ok-dialog/dv-ng-ok-dialog.component';
+import {DvNgRemoveDialogComponent} from '../../../core/component/dv-ng-remove-dialog/dv-ng-remove-dialog.component';
 import ErrorService from '../../../core/errors/service/ErrorService';
 import GesuchsperiodeRS from '../../../core/service/gesuchsperiodeRS.rest';
 import UserRS from '../../../core/service/userRS.rest';
 import ZahlungRS from '../../../core/service/zahlungRS.rest';
-import {LinkDialogController} from '../../../gesuch/dialog/LinkDialogController';
-import {OkDialogController} from '../../../gesuch/dialog/OkDialogController';
 import GemeindeRS from '../../../gesuch/service/gemeindeRS.rest';
 import GesuchRS from '../../../gesuch/service/gesuchRS.rest';
 import TSGemeinde from '../../../models/TSGemeinde';
@@ -32,8 +34,6 @@ import {ApplicationPropertyRS} from '../../service/applicationPropertyRS.rest';
 import {TestFaelleRS} from '../../service/testFaelleRS.rest';
 
 require('./testdatenView.less');
-let okDialogTempl = require('../../../gesuch/dialog/okDialogTemplate.html');
-let linkDialogTempl = require('../../../gesuch/dialog/linkDialogTemplate.html');
 
 @Component({
     selector: 'testdaten-view',
@@ -59,11 +59,12 @@ export class TestdatenViewComponent implements OnInit {
 
     devMode: boolean;
 
-    constructor(testFaelleRS: TestFaelleRS, private dvDialog: DvDialog, private userRS: UserRS,
+    constructor(testFaelleRS: TestFaelleRS, private userRS: UserRS,
                 private errorService: ErrorService, private gesuchsperiodeRS: GesuchsperiodeRS,
                 private zahlungRS: ZahlungRS, private applicationPropertyRS: ApplicationPropertyRS,
-                private gesuchRS: GesuchRS, private gemeindeRS: GemeindeRS) {
-        // todo remove dependency of DvDialog. Use MatDialog instead
+                private gesuchRS: GesuchRS, private gemeindeRS: GemeindeRS,
+                private dialog: MatDialog) {
+
         this.testFaelleRS = testFaelleRS;
     }
 
@@ -82,7 +83,7 @@ export class TestdatenViewComponent implements OnInit {
         });
     }
 
-    public createTestFallType(testFall: string): IPromise<any> {
+    public createTestFallType(testFall: string): void {
         let bestaetigt: boolean = false;
         let verfuegen: boolean = false;
         if (this.creationType === 'warten') {
@@ -98,35 +99,21 @@ export class TestdatenViewComponent implements OnInit {
             verfuegen = true;
         }
         if (this.selectedBesitzer) {
-            return this.createTestFallGS(testFall, this.selectedGesuchsperiode.id, this.selectedGemeinde.id, bestaetigt, verfuegen, this.selectedBesitzer.username);
+            this.createTestFallGS(testFall, this.selectedGesuchsperiode.id, this.selectedGemeinde.id, bestaetigt, verfuegen, this.selectedBesitzer.username);
         } else {
-            return this.createTestFall(testFall, this.selectedGesuchsperiode.id, this.selectedGemeinde.id, bestaetigt, verfuegen);
+            this.createTestFall(testFall, this.selectedGesuchsperiode.id, this.selectedGemeinde.id, bestaetigt, verfuegen);
         }
     }
 
-    private createTestFall(testFall: string, gesuchsperiodeId: string, gemeindeId: string, bestaetigt: boolean, verfuegen: boolean): IPromise<any> {
-        return this.testFaelleRS.createTestFall(testFall, gesuchsperiodeId, gemeindeId, bestaetigt, verfuegen).then((response) => {
-            //einfach die letzten 36 zeichen der response als uuid betrachten, hacky ist aber nur fuer uns intern
-            let uuidPartOfString = response.data ? response.data.slice(-36) : '';
-            return this.dvDialog.showDialog(linkDialogTempl, LinkDialogController, {
-                title: response.data,
-                link: '#/gesuch/fall/false///' + uuidPartOfString + '/', //nicht alle Parameter werden benoetigt, deswegen sind sie leer
-            }).then(() => {
-                //do nothing
-            });
+    private createTestFall(testFall: string, gesuchsperiodeId: string, gemeindeId: string, bestaetigt: boolean, verfuegen: boolean): void {
+        this.testFaelleRS.createTestFall(testFall, gesuchsperiodeId, gemeindeId, bestaetigt, verfuegen).then((response) => {
+            this.createLinkDialog(response);
         });
     }
 
-    private createTestFallGS(testFall: string, gesuchsperiodeId: string, gemeindeId: string, bestaetigt: boolean, verfuegen: boolean, username: string): IPromise<any> {
-        return this.testFaelleRS.createTestFallGS(testFall, gesuchsperiodeId, gemeindeId, bestaetigt, verfuegen, username).then((response) => {
-            //einfach die letzten 36 zeichen der response als uuid betrachten, hacky ist aber nur fuer uns intern
-            let uuidPartOfString = response.data ? response.data.slice(-36) : '';
-            return this.dvDialog.showDialog(linkDialogTempl, LinkDialogController, {
-                title: response.data,
-                link: '#/gesuch/fall/false///' + uuidPartOfString + '/', //nicht alle Parameter werden benoetigt, deswegen sind sie leer
-            }).then(() => {
-                //do nothing
-            });
+    private createTestFallGS(testFall: string, gesuchsperiodeId: string, gemeindeId: string, bestaetigt: boolean, verfuegen: boolean, username: string): void {
+        this.testFaelleRS.createTestFallGS(testFall, gesuchsperiodeId, gemeindeId, bestaetigt, verfuegen, username).then((response) => {
+            this.createLinkDialog(response);
         });
     }
 
@@ -143,44 +130,28 @@ export class TestdatenViewComponent implements OnInit {
     }
 
     public mutiereFallHeirat(): IPromise<any> {
-        return this.testFaelleRS.mutiereFallHeirat(this.dossierid, this.selectedGesuchsperiode.id,
-            this.eingangsdatum, this.ereignisdatum).then((response) => {
-            return this.dvDialog.showDialog(okDialogTempl, OkDialogController, {
-                title: response.data
-            }).then(() => {
-                //do nothing
+        return this.testFaelleRS.mutiereFallHeirat(this.dossierid, this.selectedGesuchsperiode.id, this.eingangsdatum, this.ereignisdatum)
+            .then((response) => {
+                this.createAndOpenOkDialog(response.data);
             });
-        });
     }
 
     public mutiereFallScheidung(): IPromise<any> {
-        return this.testFaelleRS.mutiereFallScheidung(this.dossierid, this.selectedGesuchsperiode.id,
-            this.eingangsdatum, this.ereignisdatum).then((respone) => {
-            return this.dvDialog.showDialog(okDialogTempl, OkDialogController, {
-                title: respone.data
-            }).then(() => {
-                //do nothing
+        return this.testFaelleRS.mutiereFallScheidung(this.dossierid, this.selectedGesuchsperiode.id, this.eingangsdatum, this.ereignisdatum)
+            .then((response) => {
+                this.createAndOpenOkDialog(response.data);
             });
-        });
     }
 
     public resetSchulungsdaten(): IPromise<any> {
         return this.testFaelleRS.resetSchulungsdaten().then((response) => {
-            return this.dvDialog.showDialog(okDialogTempl, OkDialogController, {
-                title: response.data
-            }).then(() => {
-                //do nothing
-            });
+            this.createAndOpenOkDialog(response.data);
         });
     }
 
     public deleteSchulungsdaten(): IPromise<any> {
         return this.testFaelleRS.deleteSchulungsdaten().then((response) => {
-            return this.dvDialog.showDialog(okDialogTempl, OkDialogController, {
-                title: response.data
-            }).then(() => {
-                //do nothing
-            });
+            this.createAndOpenOkDialog(response.data);
         });
     }
 
@@ -189,26 +160,65 @@ export class TestdatenViewComponent implements OnInit {
     }
 
     public deleteAllZahlungsauftraege(): void {
-        this.dvDialog.showDialog(okDialogTempl, OkDialogController, {
-            deleteText: 'ZAHLUNG_LOESCHEN_DIALOG_TEXT',
-            title: 'ZAHLUNG_LOESCHEN_DIALOG_TITLE',
-            parentController: undefined,
-            elementID: undefined
-        })
-            .then(() => {   //User confirmed removal
-                this.zahlungRS.deleteAllZahlungsauftraege();
+        this.createAndOpenRemoveDialog('ZAHLUNG_LOESCHEN_DIALOG_TITLE', 'ZAHLUNG_LOESCHEN_DIALOG_TEXT')
+            .subscribe((acceptedByUser) => {
+                if (acceptedByUser) {
+                    this.zahlungRS.deleteAllZahlungsauftraege();
+                }
             });
     }
 
     public gesuchVerfuegen(): void {
-        this.dvDialog.showDialog(okDialogTempl, OkDialogController, {
-            deleteText: 'GESUCH_VERFUEGEN_DIALOG_TEXT',
-            title: 'GESUCH_VERFUEGEN_DIALOG_TITLE',
-            parentController: undefined,
-            elementID: undefined
-        })
-            .then(() => {   //User confirmed removal
-                this.gesuchRS.gesuchVerfuegen(this.verfuegenGesuchid);
+        this.createAndOpenRemoveDialog('GESUCH_VERFUEGEN_DIALOG_TITLE', 'GESUCH_VERFUEGEN_DIALOG_TEXT')
+            .subscribe((acceptedByUser) => {
+                if (acceptedByUser) {
+                    this.gesuchRS.gesuchVerfuegen(this.verfuegenGesuchid);
+                }
             });
+    }
+
+    private createAndOpenOkDialog(title: string): void {
+        const dialogConfig = this.createDefaultMatDialogConfig();
+        dialogConfig.data = {
+            title: title,
+        };
+
+        this.dialog.open(DvNgOkDialogComponent, dialogConfig).afterClosed();
+    }
+
+    private createAndOpenRemoveDialog(title: string, text: string): Observable<boolean> {
+        const dialogConfig = this.createDefaultMatDialogConfig();
+        dialogConfig.data = {
+            title: title,
+            text: text,
+        };
+
+        return this.dialog.open(DvNgRemoveDialogComponent, dialogConfig).afterClosed();
+    }
+
+    private createLinkDialog(response: any) {
+        //einfach die letzten 36 zeichen der response als uuid betrachten, hacky ist aber nur fuer uns intern
+        let uuidPartOfString = response.data ? response.data.slice(-36) : '';
+        this.createAndOpenLinkDialog(
+            response.data,
+            '#/gesuch/fall/false///' + uuidPartOfString + '/', //nicht alle Parameter werden benoetigt, deswegen sind sie leer
+        );
+    }
+
+    private createAndOpenLinkDialog(title: string, link: string): Observable<boolean> {
+        const dialogConfig = this.createDefaultMatDialogConfig();
+        dialogConfig.data = {
+            title: title,
+            link: link,
+        };
+
+        return this.dialog.open(DvNgLinkDialogComponent, dialogConfig).afterClosed();
+    }
+
+    private createDefaultMatDialogConfig() {
+        const dialogConfig = new MatDialogConfig();
+        dialogConfig.disableClose = false; // dialog is canceled by clicking outside
+        dialogConfig.autoFocus = true;
+        return dialogConfig;
     }
 }
